+++ src/Grids/extruded.jl
@@ -28,7 +28,7 @@ uses `Geometry.DeepSphericalGlobalGeometry` when `deep = true` and
 `Geometry.ShallowSphericalGlobalGeometry` otherwise. Construction is memoized in
 `Cache.OBJECT_CACHE`.
 """
-mutable struct ExtrudedFiniteDifferenceGrid{
+struct ExtrudedFiniteDifferenceGrid{
     H <: AbstractGrid,
     V <: FiniteDifferenceGrid,
     A <: HypsographyAdaption,
@@ -160,29 +160,6 @@ global_geometry(grid::AbstractExtrudedFiniteDifferenceGrid) =
 
 quadrature_style(grid::ExtrudedFiniteDifferenceGrid) =
     quadrature_style(grid.horizontal_grid)
-
-
-## GPU compatibility
-struct DeviceExtrudedFiniteDifferenceGrid{VT, Q, GG, CLG, FLG} <:
-       AbstractExtrudedFiniteDifferenceGrid
-    vertical_topology::VT
-    quadrature_style::Q
-    global_geometry::GG
-    center_local_geometry::CLG
-    face_local_geometry::FLG
-end
-
-ClimaComms.device(::DeviceExtrudedFiniteDifferenceGrid) = DeviceSideDevice()
-ClimaComms.context(::DeviceExtrudedFiniteDifferenceGrid) = DeviceSideContext()
-
-local_geometry_type(
-    ::Type{DeviceExtrudedFiniteDifferenceGrid{VT, Q, GG, CLG, FLG}},
-) where {VT, Q, GG, CLG, FLG} = eltype(CLG) # calls eltype from DataLayouts
-
-quadrature_style(grid::DeviceExtrudedFiniteDifferenceGrid) =
-    grid.quadrature_style
-vertical_topology(grid::DeviceExtrudedFiniteDifferenceGrid) =
-    grid.vertical_topology
 
 ## aliases